Moved the entries date logic to the entries helper
